requirements.txt
Django==5.0.2
python-dotenv==1.0.1
openai==0.28.1
git+https://github.com/openai/whisper.git
pyannote.audio==3.1.1
torch==2.2.0
transformers==4.38.2
numpy==1.26.4
python-magic==0.4.27
django-cors-headers==4.3.1
djangorestframework==3.14.0
soundfile==0.12.1
ffmpeg-python==0.2.0 